pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyvmess"
description = 'a naive implementation to parse raw vmess package'
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = [
  "vmess"
]
authors = [
  { name = "Letu Ren", email = "fantasquex@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "pycryptodome",
  "fnvhash"
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/fantasquex/pyvmess#readme"
Issues = "https://github.com/fantasquex/pyvmess/issues"
Source = "https://github.com/fantasquex/pyvmess"

[tool.black]
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.pylint.format]
max-line-length = "88"

[tool.hatch.version]
path = "pyvmess/__about__.py"

[tool.hatch.build.targets.sdist]
[tool.hatch.build.targets.wheel]

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=pyvmess --cov=tests"
no-cov = "cov --no-cov"

[[tool.hatch.envs.test.matrix]]
python = ["310"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "pyvmess/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
